requirements.txt
configparser==3.5.0
Django==1.10.6
enum34==1.1.6
image==1.5.5
imagesize==0.7.1
meld3==1.0.2
MySQL-python==1.2.5
numpy==1.11.0
Pillow==3.3.1
prometheus-client==0.0.19
psycopg2==2.7
pycrypto==2.6.1
python-apt==1.1.0b1
python-dateutil==2.5.3
redis==2.10.5
scipy==0.18.1
six==1.10.0
subprocess32==3.2.7
supervisor==3.2.0